Implement AddTechnologyToProject and minor bug fixing

+++ allocation-app/src/main/java/ro/fortech/allocation/assignments/model/Assignment.java
@@ -40,6 +40,6 @@ public class Assignment {
     private String projectPosition;
 
     @Column(nullable = false, name = "hours")
-    private Integer allocationHours;
+    private Double allocationHours;
 
 }
